Processing FIBS availability messages
The Fibs object now has a State object that currently represents the user's details and the details of all the players that have been seen on the system (online and offline). The details in state are updated by asynchronously processing the messages sent by FIBS. The first set of the CLIP standard is supported, up to #8 (Logout).

Signed-off-by: Stewart M. Johnson <stewart@bolidian.com>