Added Console#status
This method updates the console to add status text to the reverse-video line at the top of the console. The text is obtained from the Status module which defines constants for the available status messages. The Status module uses I18n internally to get the actual text (default values in en.yml).

Signed-off-by: Stewart M. Johnson <stewart@bolidian.com>